user-service

+++ user-service/app/use_cases/user/delete_user_by_id.py
@@ -1,5 +1,6 @@
 from uuid import UUID
 from typing import Optional
+
 from attr import dataclass
 
 from domain.base import IUnitOfWork, BaseUseCase
@@ -7,10 +8,15 @@ from domain.entities.user import User
 
 
 @dataclass(slots=True)
-class DeleteUserById(BaseUseCase):
+class DeleteUserByIdUseCase(BaseUseCase):
     uow: IUnitOfWork
 
     async def __call__(self, user_id: UUID) -> Optional[User]:
         async with self.uow:
-            result = await self.uow.user_repo.delete_by_id(user_id)
-            return result
+            deleted_user = await self.uow.user_repo.get_by_id(user_id)
+            if not deleted_user:
+                return None
+
+            await self.uow.user_repo.delete_by_id(user_id)
+            await self.uow.commit()
+            return deleted_user
